applied requested comments, didnt fix conflicts yet
