+++ pubspec.yaml
@@ -19,7 +19,6 @@ version: 1.0.0+1
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
-#  sdk: ">=2.11.0 <3.0.0"
 
 
 # Dependencies specify other packages that your package needs in order to work.
@@ -33,11 +32,9 @@ dependencies:
     sdk: flutter
 
   http: "0.13.3"
-#  signalr_client: any
   flutter_guid: 0.2.0
   meta: ^1.7.0
   logging: ^1.0.2
-#  singalr_flutter: ^0.1.2
   signalr_netcore: 1.2.3
